# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Hello World App
# during dev, should point to your local machine
host: localhost:10010
# basePath prefixes all resource paths
basePath: /
#
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  /users/login:
    # binds a127 app logic to a route
    x-swagger-router-controller: users
    post:
      description: User Authenticate
      # used as the method name of the controller
      operationId: authenticateUser
      consumes:
        - "application/json"
      parameters:
        - name: body
          in: body
          required: true
          schema:
            type: object
            required:
              - user_email
              - user_pass
            properties:
              user_email:
                type: string
                format: email
              user_pass:
                type: string
      responses:
        "200":
          description: Success
          schema:
            type: object
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorSchema"
  /users/create:
    # binds a127 app logic to a route
    x-swagger-router-controller: users
    post:
      description: User Sign up
      # used as the method name of the controller
      operationId: createUser
      consumes:
        - multipart/form-data
      parameters:
        - in: formData
          name: profile_img
          type: file
        - in: formData
          name: user_name
          type: string
          required: true
        - in: formData
          name: user_email
          type: string
          format: email
          pattern: (?:[a-z0-9!#$%&'*+/=?^_`{|}~-]+(?:\.[a-z0-9!#$%&'*+/=?^_`{|}~-]+)*|"(?:[\x01-\x08\x0b\x0c\x0e-\x1f\x21\x23-\x5b\x5d-\x7f]|\\[\x01-\x09\x0b\x0c\x0e-\x7f])*")@(?:(?:[a-z0-9](?:[a-z0-9-]*[a-z0-9])?\.)+[a-z0-9](?:[a-z0-9-]*[a-z0-9])?|\[(?:(?:25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)\.){3}(?:25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?|[a-z0-9-]*[a-z0-9]:(?:[\x01-\x08\x0b\x0c\x0e-\x1f\x21-\x5a\x53-\x7f]|\\[\x01-\x09\x0b\x0c\x0e-\x7f])+)\])
          required: true
        - in: formData
          name: user_pass
          type: string
          required: true
      responses:
        "200":
          description: Success
          schema:
            type: object
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorSchema"

  /users/{userId}/tweets:
    # binds a127 app logic to a route
    x-swagger-router-controller: users
    get:
      description: Get Tweets made by the user
      # used as the method name of the controller
      operationId: userTweets
      parameters:
        - name: userId
          in: path
          type: integer
          required: true
      responses:
        "200":
          description: Success
          schema:
            type: object
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorSchema"

  /users/{userId}/details:
    # binds a127 app logic to a route
    x-swagger-router-controller: users
    get:
      description: Returns user details
      # used as the method name of the controller
      operationId: userDetails
      parameters:
        - name: userId
          in: path
          type: integer
          required: true
      responses:
        "200":
          description: Success
          schema:
            type: object
            properties:
              data:
                # a pointer to a definition
                $ref: "#/definitions/UserDetailsSchema"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorSchema"
  /users/{userId}/media:
    # binds a127 app logic to a route
    x-swagger-router-controller: users
    get:
      description: Returns user details
      # used as the method name of the controller
      operationId: userMedia
      parameters:
        - name: userId
          in: path
          type: integer
          required: true
      responses:
        "200":
          description: Success
          schema:
            type: object
            properties:
              data:
                type: array
                items:
                  # a pointer to a definition
                  $ref: "#/definitions/UserMediaSchema"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorSchema"
  /users/{userId}/followers/suggestions:
    # binds a127 app logic to a route
    x-swagger-router-controller: users
    get:
      description: Fetch suggestions of people the user can follow
      # used as the method name of the controller
      operationId: userFollowersSuggestions
      parameters:
        - name: userId
          in: path
          type: integer
          required: true
      responses:
        "200":
          description: Success
          schema:
            type: object
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorSchema"
  /users/{userId}/following:
    # binds a127 app logic to a route
    x-swagger-router-controller: users
    get:
      description: Fetch suggestions of people the user can follow
      # used as the method name of the controller
      operationId: userFollowing
      parameters:
        - name: userId
          in: path
          type: integer
          required: true
      responses:
        "200":
          description: Success
          schema:
            type: object
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorSchema"
  /users/{userId}/following/tweets:
    # binds a127 app logic to a route
    x-swagger-router-controller: users
    get:
      description: Get Tweets from following profiles
      # used as the method name of the controller
      operationId: userFollowingTweets
      parameters:
        - name: userId
          in: path
          type: integer
          required: true
      responses:
        "200":
          description: Success
          schema:
            type: object
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorSchema"
  /users/{userId}/followers:
    # binds a127 app logic to a route
    x-swagger-router-controller: users
    get:
      description: Get details of user's followers
      # used as the method name of the controller
      operationId: userFollowers
      parameters:
        - name: userId
          in: path
          type: integer
          required: true
        - name: query
          in: query
          type: string
          required: false
      responses:
        "200":
          description: Success
          schema:
            type: object
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorSchema"

  /swagger:
    x-swagger-pipe: swagger_raw
# complex objects have schema definitions
definitions:
  UserDetailsSchema:
    properties:
      id:
        type: number
      cover_img:
        type: string
      profile_img:
        type: string
      full_name:
        type: string

  UserMediaSchema:
    type: object
    properties:
      type:
        type: string
        enum:
          - "video"
          - "image"
  ErrorSchema:
    required:
      - message
    properties:
      message:
        type: string